Code generator: rename bpt related Python files
With renamed files one can use auto generated bpt_mem_map.py without
changes.
